+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `OpenAI` settings to generate the chat messages
 - Add `OpenAIService` to generate the chat messages
 - Add `ChatProvider` to display the chat view
+- Add `OpenAIService` instance to `ChatProvider` to generate the chat messages
 - Add `src/assets/main.js` to handle the chat view
 - Add `src/assets/main.css` to style the chat view
 - Add `getNonce` method to `security.helper.ts` file
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `writeFile` method to `saveFile` in `filesystem.helper.ts` file
 - Improve the documentation of the extension
+- Set the `getInstance` method as static in `OpenAIService` class
 
 ### Fixed