+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-02-10
+
 ### Added
 
 - Add `OpenAI` settings to generate the chat messages
@@ -97,7 +99,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix file creation issue when creating a new file
 
-[unreleased]: https://github.com/ManuelGil/extension-starter-kit/compare/v1.2.0...HEAD
+[unreleased]: https://github.com/ManuelGil/extension-starter-kit/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/ManuelGil/extension-starter-kit/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/ManuelGil/extension-starter-kit/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/ManuelGil/extension-starter-kit/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/ManuelGil/extension-starter-kit/releases/tag/v1.0.0